third_party/WebKit/LayoutTests/inspector/sources/debugger/network-uisourcecode-provider-expected.txt

CONSOLE ERROR: line 4: Uncaught SyntaxError: Unexpected token )
CONSOLE MESSAGE: line 8: Iframe loaded
Tests NetworkUISourceCodeProvider class.



Running: testDocumentResource
Creating resource.
UISourceCode: LayoutTests/inspector/sources/debugger/resources/syntax-error.html
UISourceCode is content script: false
Highlighter type: text/html
UISourceCode content: <html>
<head>
<script>
)
</script>
</head>

<body onload="console.log('Iframe loaded');">
</body>
</html>

Running: testVMScript
Creating script.
UISourceCode: foo.js
UISourceCode is content script: false
Highlighter type: text/javascript
UISourceCode content: var foo=1;


Running: testScriptResource
Creating script resource.
UISourceCode: LayoutTests/inspector/sources/debugger/resources/script1.js
UISourceCode is content script: false
Highlighter type: text/javascript
UISourceCode content: function f1() {}

Running: testRemoveStyleSheetFromModel
Creating stylesheet resource.
Added uiSourceCode: LayoutTests/inspector/sources/debugger/resources/style1.css
Removed uiSourceCode: LayoutTests/inspector/sources/debugger/resources/style1.css
